# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM1 API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/comment/index": {
            "get": {
                "tags": [
                    "comment"
                ],
                "summary": "Список комментариев",
                "operationId": "b2fc12e929639d404b15fe50305b6448",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comment"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error server"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comment/{id}": {
            "get": {
                "tags": [
                    "comment"
                ],
                "summary": "Поиск по id",
                "operationId": "aa1c887c9daa50a8a5ed40402636912a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id комментария",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "comment"
                ],
                "summary": "Удаление по id",
                "operationId": "16b0414b58d13abadda52c826dbf25f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id комментария",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comment/by-user/{id}": {
            "get": {
                "tags": [
                    "comment"
                ],
                "summary": "Поиск по id пользователя",
                "operationId": "96198a61737fb500a9abf25ac5d16dce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comment/{id}/{instance}": {
            "post": {
                "tags": [
                    "comment"
                ],
                "summary": "Создание пользователя",
                "operationId": "1f9785e606352442eccb73afec3096d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id модели",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "instance",
                        "in": "path",
                        "description": "имя модели. deal/client",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateCommentRequest"
                },
                "responses": {
                    "201": {
                        "description": "Comment resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comment/update/{id}": {
            "patch": {
                "tags": [
                    "comment"
                ],
                "summary": "Обновление по id",
                "operationId": "0b03cea92470f214eec2378cdc64caa9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id комментерия",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateCommentRequest"
                },
                "responses": {
                    "200": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Comment": {
                "required": [
                    "body"
                ],
                "properties": {
                    "body": {
                        "type": "string",
                        "example": "test body comment"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
...
            "CreateCommentRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "user_id",
                                "body"
                            ],
                            "properties": {
                                "user_id": {
                                    "type": "integer",
                                    "example": "1"
                                },
                                "body": {
                                    "type": "string",
                                    "example": "example body comment"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateCommentRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [],
                            "properties": {
                                "user_id": {
                                    "type": "integer",
                                    "example": "1",
                                    "nullable": true
                                },
                                "body": {
                                    "type": "string",
                                    "example": "example body comment",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}